refactor: Reorganize file structure to align with atomic design principles
Optimize file paths and directory structure
