+++ kakao-shopping/src/components/atoms/MainSkeleton.jsx
