+++ kakao-shopping/src/components/atoms/StarRating.jsx
